# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rayito de Sol API",
        "description": "Documentación de la API para Rayito de Sol",
        "contact": {
            "email": "sergioramosalarcon@rayitodesol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/admin/stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Estadísticas del sistema para administradores",
                "operationId": "8bc93127d5d5f4382be3d67a3c10694a",
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_users": {
                                            "type": "integer"
                                        },
                                        "active_users": {
                                            "type": "integer"
                                        },
                                        "admin_users": {
                                            "type": "integer"
                                        },
                                        "recent_registrations": {
                                            "type": "integer"
                                        },
                                        "total_properties": {
                                            "type": "integer"
                                        },
                                        "total_reservations": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}